extract Widget base class to separate lib/widget.py
