+++ selfdrive/ui/layouts/network.py
@@ -1,5 +1,5 @@
 import pyray as rl
-from openpilot.system.ui.lib.application import Widget
+from openpilot.system.ui.lib.widget import Widget
 from openpilot.system.ui.lib.wifi_manager import WifiManagerWrapper
 from openpilot.system.ui.widgets.network import WifiManagerUI
