+++ selfdrive/ui/layouts/settings/toggles.py
@@ -1,5 +1,5 @@
-from openpilot.system.ui.lib.application import Widget
 from openpilot.system.ui.lib.list_view import ListView, toggle_item
+from openpilot.system.ui.lib.widget import Widget
 from openpilot.common.params import Params
 
 # Description constants
