+++ selfdrive/ui/layouts/sidebar.py
@@ -4,8 +4,10 @@ from dataclasses import dataclass
 from collections.abc import Callable
 from cereal import log
 from openpilot.selfdrive.ui.ui_state import ui_state
-from openpilot.system.ui.lib.application import gui_app, FontWeight, Widget
+from openpilot.system.ui.lib.application import gui_app, FontWeight
 from openpilot.system.ui.lib.text_measure import measure_text_cached
+from openpilot.system.ui.lib.widget import Widget
+
 
 SIDEBAR_WIDTH = 300
 METRIC_HEIGHT = 126
