+++ selfdrive/ui/onroad/cameraview.py
@@ -5,8 +5,9 @@ import pyray as rl
 from openpilot.system.hardware import TICI
 from msgq.visionipc import VisionIpcClient, VisionStreamType, VisionBuf
 from openpilot.common.swaglog import cloudlog
-from openpilot.system.ui.lib.application import gui_app, Widget
+from openpilot.system.ui.lib.application import gui_app
 from openpilot.system.ui.lib.egl import init_egl, create_egl_image, destroy_egl_image, bind_egl_image_to_texture, EGLImage
+from openpilot.system.ui.lib.widget import Widget
 
 CONNECTION_RETRY_INTERVAL = 0.2  # seconds between connection attempts
