+++ selfdrive/ui/onroad/hud_renderer.py
@@ -3,9 +3,11 @@ from dataclasses import dataclass
 from cereal.messaging import SubMaster
 from openpilot.selfdrive.ui.ui_state import ui_state, UIStatus
 from openpilot.selfdrive.ui.onroad.exp_button import ExpButton
-from openpilot.system.ui.lib.application import gui_app, FontWeight, Widget
+from openpilot.system.ui.lib.application import gui_app, FontWeight
 from openpilot.system.ui.lib.text_measure import measure_text_cached
 from openpilot.common.conversions import Conversions as CV
+from openpilot.system.ui.lib.widget import Widget
+
 
 # Constants
 SET_SPEED_NA = 255
