+++ selfdrive/ui/widgets/exp_mode_button.py
@@ -1,7 +1,7 @@
 import pyray as rl
-from openpilot.system.ui.lib.application import Widget
 from openpilot.common.params import Params
 from openpilot.system.ui.lib.application import gui_app, FontWeight
+from openpilot.system.ui.lib.widget import Widget
 
 
 class ExperimentalModeButton(Widget):
