+++ selfdrive/ui/widgets/setup.py
@@ -2,9 +2,10 @@ import pyray as rl
 from openpilot.selfdrive.ui.lib.prime_state import PrimeType
 from openpilot.selfdrive.ui.widgets.pairing_dialog import PairingDialog
 from openpilot.selfdrive.ui.ui_state import ui_state
-from openpilot.system.ui.lib.application import gui_app, Widget, FontWeight
+from openpilot.system.ui.lib.application import gui_app, FontWeight
 from openpilot.system.ui.lib.button import gui_button, ButtonStyle
 from openpilot.system.ui.lib.wrap_text import wrap_text
+from openpilot.system.ui.lib.widget import Widget
 
 
 class SetupWidget(Widget):
