+++ selfdrive/ui/widgets/ssh_key.py
@@ -5,7 +5,7 @@ import copy
 from enum import Enum
 
 from openpilot.common.params import Params
-from openpilot.system.ui.lib.application import gui_app, DialogResult, FontWeight
+from openpilot.system.ui.lib.application import gui_app, FontWeight
 from openpilot.system.ui.lib.button import gui_button, ButtonStyle
 from openpilot.system.ui.lib.list_view import (
   ItemAction,
@@ -16,6 +16,7 @@ from openpilot.system.ui.lib.list_view import (
   BUTTON_WIDTH,
 )
 from openpilot.system.ui.lib.text_measure import measure_text_cached
+from openpilot.system.ui.lib.widget import DialogResult
 from openpilot.system.ui.widgets.confirm_dialog import alert_dialog
 from openpilot.system.ui.widgets.keyboard import Keyboard
