+++ system/ui/lib/list_view.py
@@ -4,11 +4,12 @@ from dataclasses import dataclass
 from collections.abc import Callable
 from abc import ABC
 from openpilot.system.ui.lib.scroll_panel import GuiScrollPanel
-from openpilot.system.ui.lib.application import gui_app, FontWeight, Widget
+from openpilot.system.ui.lib.application import gui_app, FontWeight
 from openpilot.system.ui.lib.text_measure import measure_text_cached
 from openpilot.system.ui.lib.wrap_text import wrap_text
 from openpilot.system.ui.lib.button import gui_button, ButtonStyle
 from openpilot.system.ui.lib.toggle import Toggle, WIDTH as TOGGLE_WIDTH, HEIGHT as TOGGLE_HEIGHT
+from openpilot.system.ui.lib.widget import Widget
 
 ITEM_BASE_HEIGHT = 170
 LINE_PADDING = 40
