+++ system/ui/widgets/confirm_dialog.py
@@ -1,7 +1,9 @@
 import pyray as rl
-from openpilot.system.ui.lib.application import gui_app, DialogResult, FontWeight
+from openpilot.system.ui.lib.application import gui_app, FontWeight
 from openpilot.system.ui.lib.button import gui_button, ButtonStyle
 from openpilot.system.ui.lib.label import gui_text_box
+from openpilot.system.ui.lib.widget import DialogResult
+
 
 DIALOG_WIDTH = 1520
 DIALOG_HEIGHT = 600
