+++ system/ui/widgets/option_dialog.py
@@ -1,8 +1,10 @@
 import pyray as rl
-from openpilot.system.ui.lib.application import Widget, FontWeight
+from openpilot.system.ui.lib.application import FontWeight
 from openpilot.system.ui.lib.button import gui_button, ButtonStyle, TextAlignment
 from openpilot.system.ui.lib.label import gui_label
 from openpilot.system.ui.lib.scroll_panel import GuiScrollPanel
+from openpilot.system.ui.lib.widget import Widget
+
 
 # Constants
 MARGIN = 50
